% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{colog}
\alias{colog}
\title{Compute Logarithmic Transformation for Probabilities}
\usage{
colog(prob)
}
\arguments{
\item{prob}{A numeric vector of probabilities.}
}
\value{
A numeric vector where values less than \code{LIMIT_PROB} are replaced
by \code{LIMIT_COLOG}, and other values are transformed using \code{-log10()}.
}
\description{
This function computes a custom logarithmic transformation of probabilities.
If the probability is less than \code{LIMIT_PROB}, it returns \code{LIMIT_COLOG}.
Otherwise, it returns the negative base-10 logarithm of the probability.
}
\examples{
# Example with a single probability value
colog(0.001)

# Example with a vector of probabilities
colog(c(0.001, 0.05, 0.1))

}
