% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_panalyzer.R
\docType{data}
\name{example_panalyzer}
\alias{example_panalyzer}
\title{PAnalyzer Example Data}
\format{
A data frame with 18 rows and the following 5 columns:
\describe{
\item{peptideRef}{Character vector of peptide references.}
\item{proteinRef}{Character vector of protein references.}
\item{peptideType}{Character vector indicating the type of peptide.}
\item{proteinType}{Character vector indicating the type of protein.}
\item{groupRef}{Integer vector of protein group references.}
}
}
\source{
Data extracted from \href{https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-13-288}{PAnalyzer: A software tool for protein inference in shotgun proteomics} for demonstration purposes.
}
\usage{
example_panalyzer
}
\description{
A dataset containing the PAnalyzer example presented in the original paper.
}
\examples{
data(example_panalyzer, package = "b10prot")
plot_groups(example_panalyzer, groupRefs = 1:5)

}
\keyword{datasets}
