% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_workflow.R
\name{iwf_grouping}
\alias{iwf_grouping}
\title{Perform Protein Grouping Based on Peptide-to-Protein Relations}
\usage{
iwf_grouping(pep2prot, threshold = 0.01)
}
\arguments{
\item{pep2prot}{A data frame containing peptide-to-protein relations, including
the columns \code{peptideRef}, \code{proteinRef}, \code{qval}, and \code{isDecoy}.}

\item{threshold}{A numeric value specifying the peptide-level FDR threshold
for grouping (default is 0.01).}
}
\value{
A data frame with the inferred protein groups, with additional columns:
\itemize{
\item \code{peptideType}: Type of peptide (\code{"unique"}, \code{"discriminating"}, \code{"non-discriminating"}).
\item \code{proteinType}: Type of protein (\code{"conclusive"}, \code{"indistinguishable"}, \code{"ambiguous"}, \code{"non-conclusive"}).
\item \code{groupRef}: Group identifier for the proteins.
\item \code{shared}: The number of groups matched by each peptide.
}
}
\description{
This function builds protein groups using the peptides that pass a
specified peptide-level FDR threshold. Peptides that do not pass the threshold
are grouped separately and assigned a negative group identifier. It utilizes
the PAnalyzer algorithm to infer protein groups and assigns peptide and protein types.
}
\seealso{
\code{\link{panalyzer}} for protein grouping, \code{\link{iwf_pep2level}} for obtaining peptide-to-protein relations.
}
