% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_workflow.R
\name{iwf_load_psms}
\alias{iwf_load_psms}
\title{Load PSMs from mzIdentML Files}
\usage{
iwf_load_psms(path = ".", pattern = ".mzid", psm_score = NULL, verbose = FALSE)
}
\arguments{
\item{path}{A character string specifying the directory where the mzIdentML files are located.
Default is the current working directory (\code{"."}).}

\item{pattern}{A character string representing the file pattern to search for. Default is \code{".mzid"},
which targets mzIdentML files.}

\item{psm_score}{An optional character string specifying the column name to be used as the PSM score.
If \code{NULL}, the last column in the loaded data will be used as the PSM score.}

\item{verbose}{Logical, whether to print progress messages (default TRUE).}
}
\value{
A \code{data.frame} containing the combined PSM data, with added columns for the PSM score
(\code{psmScore}) and the protein reference (\code{proteinRef}).
}
\description{
This function loads Peptide-to-Spectrum Matches (PSMs) from mzIdentML files in a specified directory,
combines them, and adds a standard column for the PSM score and protein reference.
}
