% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panalyzer.R
\name{plot_groups}
\alias{plot_groups}
\title{Plot Protein Groups Composition}
\usage{
plot_groups(panalyzer, groupRefs)
}
\arguments{
\item{panalyzer}{A data frame containing the results from the \code{panalyzer} function. The data frame
must include columns such as \code{groupRef}, \code{peptideRef}, \code{peptideType}, \code{proteinRef} and \code{proteinType}.}

\item{groupRefs}{A vector of group reference identifiers (\code{groupRef}) representing the protein
groups to be visualized.}
}
\value{
A ggplot object visualizing the selected protein groups and their associated peptides.
}
\description{
Visualizes the composition of selected protein groups based on their associated peptides.
The function filters the specified protein groups, arranges them, and displays a plot showing
the peptides associated with each protein in the group. Peptides are colored according to
their classification as unique, discriminating, non-discriminating, or non-significant.
}
\examples{
data(example_panalyzer, package = "b10prot")
plot_groups(example_panalyzer, groupRefs = 1:5)

}
\seealso{
\itemize{
\item \href{https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-13-288}{PAnalyzer: A software tool for protein inference in shotgun proteomics} for an example plot.
}
}
