% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{refined_fdr}
\alias{refined_fdr}
\title{Refined False Discovery Rate (FDR) Calculation}
\usage{
refined_fdr(data, levelRef, score, lower_better = TRUE, affix = "_REVERSED")
}
\arguments{
\item{data}{A data frame containing the identification data, including
columns for the reference level, a score, and whether each identification
is a decoy (\code{isDecoy}).}

\item{levelRef}{The column name containing the reference level for each
identification (e.g., protein or gene reference). This should be an unquoted
column name.}

\item{score}{The column name of the score used to rank the identifications.
This should be an unquoted column name.}

\item{lower_better}{A logical value indicating whether lower scores are
better (default is \code{TRUE}).}

\item{affix}{A string indicating the suffix of prefix used to identify decoy
entries in the reference level column. Default is \code{"_REVERSED"}.}
}
\value{
A data frame with the original data and additional columns for the
refined FDR estimates:
\describe{
\item{FDRn}{Normal FDR estimation as cumulative minimum (q-value).}
\item{FDRp}{Picked FDR estimation as cumulative minimum (q-value).}
\item{FDRr}{Refined FDR estimationas cumulative minimum (q-value).}
\item{to}{Target-only identifications count.}
\item{do}{Decoy-only identifications count.}
\item{td}{Count of identifications with the same target and decoy scores.}
\item{tb}{Target-best identifications count.}
\item{db}{Decoy-best identifications count.}
}
}
\description{
This function computes refined False Discovery Rate (FDR) estimates using a
competitive approach between target and decoy identifications. It provides
three types of FDR calculations: FDRn, FDRp, and FDRr, which adjust for
different competitive scenarios between targets and decoys.
}
\examples{
# Example usage with a sample dataset
sample_data <- data.frame(
  proteinRef = c("P1", "P1_REVERSED", "P2", "P3", "P3_REVERSED"),
  score = c(0.1, 0.2, 0.3, 0.5, 0.4),
  isDecoy = c(FALSE, TRUE, FALSE, FALSE, TRUE)
)
refined_fdr(sample_data, levelRef = proteinRef, score = score, lower_better = TRUE)

}
\seealso{
\itemize{
\item \href{https://pubs.acs.org/doi/10.1021/acs.jproteome.9b00819}{Protein Probability Model for High-Throughput Protein Identification by Mass Spectrometry-Based Proteomics} for more information on the refined FDR estimation.
}
}
