% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{target_decoy_approach}
\alias{target_decoy_approach}
\title{Target-Decoy Approach for FDR Estimation}
\usage{
target_decoy_approach(data, score, lower_better = TRUE, unbiased = FALSE)
}
\arguments{
\item{data}{A data frame containing the identification data. It must
include a logical column \code{isDecoy} indicating whether each row corresponds
to a decoy identification.}

\item{score}{The column name of the score used to rank identifications.
This should be an unquoted column name.}

\item{lower_better}{A logical value indicating whether lower scores are
better (default is \code{TRUE}).}

\item{unbiased}{When set to \code{TRUE}, one is added to the decoy and target
counts before FDR calculation. This adjustment can reduce bias in small datasets.}
}
\value{
A data frame with the original data and additional columns:
\describe{
\item{decoys}{The cumulative number of decoys up to each identification.}
\item{targets}{The cumulative number of targets up to each identification.}
\item{pval}{The p-value estimated using the target-decoy approach.}
\item{LP}{The local confidence score, computed using the \code{colog()} function.}
\item{FDR}{The false discovery rate (FDR) for each score threshold.}
\item{qval}{The cumulative minimum FDR (q-value).}
}
}
\description{
This function applies the Target-Decoy Approach (TDA) to estimate the
False Discovery Rate (FDR) based on a scoring metric for the potential
identifications. It calculates the p-value, local confidence score (LP),
and q-value (FDR) for each identification.
}
\examples{
# Example usage with a sample dataset
sample_data <- data.frame(
  score = c(0.01, 0.02, 0.03, 0.04, 0.04, 0.05, 0.06, 0.07),
  isDecoy = c(FALSE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE, TRUE)
)
target_decoy_approach(sample_data, score, lower_better = TRUE)

}
