% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{wf_pep2level}
\alias{wf_pep2level}
\title{Map Peptides to a Specified Reference Level}
\usage{
wf_pep2level(psms, levelRef)
}
\arguments{
\item{psms}{A \code{data.frame} containing PSM data, including \code{peptideRef} (peptide identifier) and the reference level specified by \code{levelRef}.}

\item{levelRef}{The column name in \code{psms} that contains the reference level to which peptides should be mapped.}
}
\value{
A \code{data.frame} containing:
\item{\code{peptideRef}}{The peptide identifier.}
\item{\code{levelRef}}{The reference level identifier.}
\item{\code{shared}}{The number of reference entities matched by each peptide.}
}
\description{
This function maps peptides to a specified reference level (e.g., protein, gene, or protein group)
based on Peptide-to-Spectrum Matches (PSMs). It also calculates how many reference entities
(e.g., proteins, genes, or groups) are matched by each peptide.
}
