% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{lpg}
\alias{lpg}
\title{LP Gamma (LPG) Metrics Calculation}
\usage{
lpg(data, levelRef, threshold = 0.01)
}
\arguments{
\item{data}{A data frame containing identification data, including
peptide-related columns for coLogarithm of Probability (\code{LP}), q-values
(\code{qval}), and a logical column \code{isDecoy}.}

\item{levelRef}{The column name of the reference level to group by, such as
a protein or a gene identifier. This should be an unquoted column name.}

\item{threshold}{A numeric value representing the FDR threshold for
peptide-level q-values (default is \code{0.01}).}
}
\value{
A data frame containing the calculated LPG metrics:
\describe{
\item{isDecoy}{Indicates whether the group contains any decoy identification.}
\item{n}{The total number of peptide identifications for the group.}
\item{m}{The number of peptide identifications with a q-value below the threshold.}
\item{LPM}{The maximum coLogarithm of Probability (\code{LP}) for the group.}
\item{LPS}{The sum of coLogarithm of Probability (\code{LP}) for the group.}
\item{LPF}{The sum of coLogarithm of Probability for identifications with a q-value below the threshold.}
\item{LPGM}{The LP Gamma value based on the maximum \code{LP}.}
\item{LPGS}{The LP Gamma value based on the sum of \code{LP} values.}
\item{LPGF}{The LP Gamma value based on the filtered sum of \code{LP} values for confident identifications.}
}
}
\description{
This function calculates various LP Gamma (LPG) metrics for a specified
reference level, such as a protein, a gene or a protein group. The metrics are
derived from the coLogarithm of Probability (LP) of their constituting peptides
and include probabilities derived from maximum LP, sum of LP values, and
filtered LP sums using a specified peptide-level FDR threshold.
}
\examples{
# Example usage with a sample dataset
sample_data <- data.frame(
  levelRef = c("P1", "P1", "P2", "P2", "P3"),
  LP = c(1.5, 2.0, 0.5, 1.0, 1.2),
  qval = c(0.01, 0.02, 0.005, 0.03, 0.01),
  isDecoy = c(FALSE, FALSE, TRUE, FALSE, FALSE)
)
lpg(sample_data, levelRef, threshold = 0.01)

}
\seealso{
\itemize{
\item \href{https://pubs.acs.org/doi/10.1021/acs.jproteome.9b00819}{Protein Probability Model for High-Throughput Protein Identification by Mass Spectrometry-Based Proteomics} for more information on the LPG scores.
}
}
