% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{picked_gfdr}
\alias{picked_gfdr}
\title{Picked Protein Group FDR}
\usage{
picked_gfdr(data, score, lower_better = TRUE, affix = "_REVERSED")
}
\arguments{
\item{data}{A tibble containing protein group information. Must include at
least the columns groupRef, proteinMaster, proteinRefs, isDecoy and the scoring
column provided in score.}

\item{score}{The column name of the score used to rank the identifications.
This should be an unquoted column name.}

\item{lower_better}{Logical; if TRUE, lower scores indicate better identifications (default \code{TRUE}).}

\item{affix}{String indicating the suffix/prefix used to identify decoy entries (default \code{"_REVERSED"}).}
}
\value{
A tibble containing the subset of protein groups that passed the
picked protein group filtering and the subsequent target-decoy FDR analysis.
The output will typically include the original columns from data and any
additional FDR-related metrics added by target_decoy_approach().
}
\description{
Implements the picked protein group false discovery rate (FDR) strategy as
described by The et al. (2022).
}
\details{
Protein groups are first sorted by a scoring column, and for each leading protein
any subsequent group containing it within its members is removed.

Internally, the function:
\enumerate{
\item Normalizes protein identifiers by removing the affix (decoy tag).
\item Orders protein groups according to the chosen score and direction.
\item Compares each group against subsequent groups to do the competition.
\item Retains only the winner groups.
\item Applies a target-decoy approach to estimate FDR.
}
}
\examples{
library(dplyr)
library(stringr)

df <- tibble(
  groupRef = c("a", "b", "c", "d", "e", "f_REVERSED", "f"),
  proteinMaster = c("a", "b", "c", "d", "e", "f_REVERSED", "f"),
  proteinRefs = c(
    "a;x",
    "a_REVERSED;b_REVERSED",
    "b_REVERSED;c_REVERSED;x_REVERSED",
    "d",
    "e;d",
    "f_REVERSED",
    "f"
  ),
  score = 7:1
) \%>\%
  mutate(isDecoy = str_detect(proteinRefs, "_REVERSED"))

result <- pgT(df, score, lower_better = FALSE)
glimpse(result)

}
\references{
The M, Samaras P, Kuster B, Wilhelm M.
Reanalysis of ProteomicsDB Using an Accurate, Sensitive, and Scalable False
Discovery Rate Estimation Approach for Protein Groups.
Mol Cell Proteomics. 2022 Dec;21(12):100437. \doi{10.1016/j.mcpro.2022.100437}
}
\seealso{
\code{\link[=target_decoy_approach]{target_decoy_approach()}}
}
\keyword{internal}
